SProGame

Documentation

data Player

Constructors

Client

Server

type Stage = Int

Stage is just an Int

s :: Stage >= 0

data Module

Constructors

Module

Fields

modName :: String

importance :: Double

reversingTime :: Stage

type Service = [Module]

A Service is a set of Modules

data Target

Any action targets a module or nothing (Bottom)

Constructors

Target Module

Bottom

data Outcome

Constructors

Prevented Target

NotPrevented Target

NoAttack

totalValue :: Service -> Double

score :: Player -> Service -> Outcome -> Double

The score is the utility a Player gets from the Outcome of Game on a Service

type Move = (Target, Target)

A Move is a pair of choices: (Server choice, Client choice)

type Game = [Move]

A Game is a list of Moves indexed by Stage

play :: Game -> Outcome

type Strategy = [Int]

A Strategy is a list of non negative Ints used to compute the probability distribution of Targets. For example, if Service has 3 Modules, [42, 42, 42, 42] will induce (via drawFromStrategy) a uniform distribution over each of the Modules plus Bottom (head of the list).

if (length (ss :: Service)) == n then (length (s :: Strategy)) == n + 1

nextMove :: Player -> Service -> Strategy -> StdGen -> (Target, StdGen)

nextMove is chosen randomly according to a Strategy

getStrategy

Arguments

:: Player

-> Service

-> [Target]
history of this Player moves

-> Strategy
the head of the list will be used for Bottom

Returns the strategy for a Player given the history of its moves

drawFromStrategy :: StdGen -> Strategy -> (Int, StdGen)

mkGame :: Int -> Service -> StdGen -> Game

Make a Game with a given number of Moves

main :: IO ()

A simple simulated game